+++ src/importlinter/application/ports/building.py
@@ -1,4 +1,5 @@
 import abc
+from typing import List
 
 from importlinter.domain.ports.graph import ImportGraph
 
@@ -6,6 +7,6 @@ from importlinter.domain.ports.graph import ImportGraph
 class GraphBuilder(abc.ABC):
     @abc.abstractmethod
     def build(
-        self, root_package_name: str, include_external_packages: bool = False
+        self, root_package_names: List[str], include_external_packages: bool = False
     ) -> ImportGraph:
         raise NotImplementedError