+++ tests/assets/multipleroots/rootpackagegreen/two.py
@@ -1,0 +1,3 @@
+from rootpackageblue.one import alpha
+
+from . import one