+++ CHANGELOG.rst
@@ -45,3 +45,9 @@ Changelog
 * Add https://pre-commit.com configuration.
 * Use find_shortest_chains instead of find_shortest_chain on the Grimp import graph.
 * Add Forbidden Modules contract type.
+
+latest
+------
+
+* Allow multiple root packages.
+* Make containers optional in Layers contracts.